fix(browse): keep the grid where you left it after a mod visit
- Opening a mod unmounted `Browse` — it and `ModDetail` are the two arms of one
  ternary in both the Dashboard and the overlay — so the search text, category
  pill, sort order, every "load more" page, the selection and the scroll offset
  were all thrown away, and Back refetched an unfiltered page 1.
- Move that state into a new `useModListing` hook, called by `useModBrowsing`
  above the swap. Nothing it holds unmounts, so coming back re-renders the grid
  already in memory and fires no request at all.
- `Browse` keeps only its install machinery and reads the listing from one prop;
  the scroll container records its offset and restores it in a layout effect, so
  lazy thumbnails load for the right row and there is no jump.
- Same fix in the in-game overlay, which renders the same two components. Also
  survives a trip to Library/Presets/Settings, since the state now lives above
  the whole view switch.
- Unchanged: a mod-type switch still resets the category pill and selection (and
  now the scroll) while keeping the search text; a game switch still remaps.

No DB or schema changes.
